Fixed bug with VM migration where a zone could not be successfully reprovisioned after migration
This was because the "mountpoint" ZFS property was being explicitly preserved which means the source of the ZFS property was changing from "default" to "local" and when that happens the mountpoint does not get automatically updated when the reprovision routine renames the dataset which causes the new zone root that gets cloned from the image to fail to mount since the original zone root is still mounted at the original location. To fix this the list of properties to potentially preserve is now filtered by properties with source=local.

Also added preservation of some additional ZFS properties that should be preserved if set (sharesmb, zoned, devices).